⚡️ :: period picker
